+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Crate:
 Protocol:
 - Valve:
 1. Renamed `protocol` to `protocol_version`.
+2. Renamed `version` to `game_version`.
 
 - Minecraft (both Java and Bedrock)
 1. Renamed `version_protocol` to `protocol_version`.